Rgb class refactoring
