+++ YUViewLib/src/video/frameHandler.cpp
@@ -37,7 +37,13 @@
 #include "common/functions.h"
 #include "playlistitem/playlistItem.h"
 
-// ------ Initialize the static list of frame size presets ----------
+// Activate this if you want to know when which buffer is loaded/converted to image and so on.
+#define FRAMEHANDLER_DEBUG_LOADING 0
+#if FRAMEHANDLER_DEBUG_LOADING && !NDEBUG
+#define DEBUG_FRAME qDebug
+#else
+#define DEBUG_FRAME(fmt,...) ((void)0)
+#endif
 
 class frameHandler::frameSizePresetList
 {
@@ -79,16 +85,6 @@ QStringList frameHandler::frameSizePresetList::getFormattedNames() const
 }
 
 frameHandler::frameSizePresetList frameHandler::presetFrameSizes;
-
-// ---------------- frameHandler ---------------------------------
-
-// Activate this if you want to know when which buffer is loaded/converted to image and so on.
-#define FRAMEHANDLER_DEBUG_LOADING 0
-#if FRAMEHANDLER_DEBUG_LOADING && !NDEBUG
-#define DEBUG_FRAME qDebug
-#else
-#define DEBUG_FRAME(fmt,...) ((void)0)
-#endif
 
 frameHandler::frameHandler()
 {
